pyproject.toml
[tool.poetry]
name = "gobgift"
version = "0.1.0"
description = "A simple wishlist manager"
authors = ["Kim Desrosiers <kimdesro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
django-environ = "^0.4.5"
django-autocomplete-light = "^3.3"
djangorestframework = "^3.9"
django-rest-swagger = "^2.2"
django-floppyforms = "^1.7"
django-allauth = "^0.38.0"
django-rest-auth = "^0.9.3"
django-webpack-loader = "^0.6.0"
psycopg2 = "^2.7"
pillow = "^5.4"

[tool.poetry.dev-dependencies]
pytest = "^4.1"
pytest-django = "^3.4"
pytest-cov = "^2.6"
django-debug-toolbar = "^1.11"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
